@@ -8,38 +8,39 @@
   "main": "index.js",
   "dependencies": {
     "autoprefixer": "^7.1.0",
-    "babel-core": "https://registry.npmjs.org/babel-core/-/babel-core-6.22.1.tgz",
-    "babel-plugin-module-resolver": "https://registry.npmjs.org/babel-plugin-module-resolver/-/babel-plugin-module-resolver-2.6.2.tgz",
+    "babel-core": "^6.22.1",
+    "babel-plugin-module-resolver": "^2.6.2",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-plugin-wrap-in-js": "https://registry.npmjs.org/babel-plugin-wrap-in-js/-/babel-plugin-wrap-in-js-1.1.1.tgz",
+    "babel-plugin-wrap-in-js": "^1.1.1",
+    "basic-auth": "^1.1.0",
     "bluebird": "^3.5.0",
-    "body-parser": "https://registry.npmjs.org/body-parser/-/body-parser-1.16.0.tgz",
+    "body-parser": "^1.16.0",
     "bourbon": "^4.3.4",
     "brace": "^0.10.0",
     "chroma-js": "^1.3.3",
     "classnames": "^2.2.5",
     "cookie-parser": "^1.4.3",
-    "d3": "https://registry.npmjs.org/d3/-/d3-3.5.17.tgz",
+    "cookie-session": "^2.0.0-beta.2",
+    "d3": "^3.5.17",
     "datalib": "^1.7.3",
     "dotenv": "^4.0.0",
-    "es-decorators": "https://registry.npmjs.org/es-decorators/-/es-decorators-1.0.0.tgz",
+    "es-decorators": "^1.0.0",
     "esri-leaflet": "^2.0.8",
     "express": "^4.14.0",
-    "express-session": "^1.14.2",
     "extract-text-webpack-plugin": "^2.1.0",
     "file-loader": "^0.11.1",
     "foundation-sites": "^6.3.1",
-    "glob": "https://registry.npmjs.org/glob/-/glob-7.1.1.tgz",
+    "glob": "^7.1.1",
     "image-webpack-loader": "^3.3.1",
     "isomorphic-fetch": "^2.2.1",
     "jiminy": "^0.1.1",
-    "json-loader": "https://registry.npmjs.org/json-loader/-/json-loader-0.5.4.tgz",
+    "json-loader": "^0.5.4",
     "leaflet": "^1.0.3",
-    "lodash": "https://registry.npmjs.org/lodash/-/lodash-4.17.4.tgz",
+    "lodash": "^4.17.4",
     "lusca": "^1.4.1",
     "moveto": "^1.6.1",
     "next": "^2.4.3",
-    "next-redux-wrapper": "https://registry.npmjs.org/next-redux-wrapper/-/next-redux-wrapper-1.1.2.tgz",
+    "next-redux-wrapper": "^1.1.2",
     "next-routes": "^1.0.27",
     "node-sass": "^4.5.0",
     "nodemailer": "^2.7.0",
@@ -47,26 +48,27 @@
     "now-logs": "0.0.7",
     "orm": "^3.2.3",
     "passport": "^0.3.2",
-    "passport-control-tower": "https://registry.npmjs.org/passport-control-tower/-/passport-control-tower-0.1.1.tgz",
+    "passport-control-tower": "^0.1.1",
     "passport-facebook": "^2.1.1",
     "passport-google-oauth": "^1.0.0",
     "passport-twitter": "^1.0.4",
-    "postcss-easy-import": "https://registry.npmjs.org/postcss-easy-import/-/postcss-easy-import-2.0.0.tgz",
+    "postcss-easy-import": "^2.0.0",
     "postcss-loader": "^2.0.5",
-    "prop-types": "https://registry.npmjs.org/prop-types/-/prop-types-15.5.7.tgz",
+    "prop-types": "^15.5.7",
     "raw-loader": "^0.5.1",
     "rc-pagination": "^1.8.8",
-    "react": "^15.4.2",
+    "react": "^15.5.4",
     "react-ace": "^4.3.0",
     "react-dnd": "^2.4.0",
     "react-dnd-html5-backend": "^2.4.1",
-    "react-dom": "^15.4.2",
+    "react-dom": "^15.5.4",
     "react-dropzone": "^3.13.1",
     "react-input-range": "^1.1.4",
     "react-redux": "^4.4.6",
     "react-select": "^1.0.0-rc.4",
     "react-slick": "^0.14.11",
     "react-sortable-hoc": "^0.6.3",
+    "react-sticky": "^6.0.1",
     "react-tether": "^0.5.7",
     "react-tokeninput": "^2.4.0",
     "react-upload-file": "^2.0.0-beta.6",
@@ -80,7 +82,7 @@
     "url-loader": "^0.5.8",
     "uuid": "^3.0.1",
     "vega": "^2.6.0",
-    "whatwg-fetch": "https://registry.npmjs.org/whatwg-fetch/-/whatwg-fetch-2.0.2.tgz"
+    "whatwg-fetch": "^2.0.2"
   },
   "devDependencies": {
     "babel-eslint": "^7.2.3",
