@@ -1,10 +1,15 @@
 import React from 'react';
+import PropTypes from 'prop-types';
+import classnames from 'classnames';
+import debounce from 'lodash/debounce';
+
+// Next
 import { Link } from 'routes';
 
 // Components
-import Logo from 'components/app/layout/Logo';
-import Menu from 'components/ui/Menu';
-import TetherComponent from 'react-tether';
+import HeaderDropdownData from 'components/app/layout/header/HeaderDropdownData';
+import HeaderDropdownAbout from 'components/app/layout/header/HeaderDropdownAbout';
+import HeaderUser from 'components/app/layout/header/HeaderUser';
 
 class Header extends React.Component {
 
@@ -12,170 +17,84 @@ class Header extends React.Component {
     super(props);
 
     this.state = {
-      dataDropdownActive: false,
-      aboutDropdownActive: false
+      dataActive: false,
+      aboutActive: false
     };
 
     this.listeners = {};
 
     // BINDINGS
-    this.toggleDataDropdown = this.toggleDataDropdown.bind(this);
-    this.onScreenClick = this.onScreenClick.bind(this);
+    this.toggleDropdown = debounce(this.toggleDropdown.bind(this), 50);
   }
 
-  componentWillUnmount() {
-    window.removeEventListener('click', this.onScreenClick);
-  }
-
-  onScreenClick(e) {
-
-    const el = document.querySelector('.c-tooltip');
-    const clickOutside = el && el.contains && !el.contains(e.target);
-    const isDataBtn = this.dataDropdownBtn.contains(e.target);
-    const isAboutBtn = this.aboutDropdownBtn.contains(e.target);
-
-    if (clickOutside) {
-      if (!isDataBtn) this.toggleDataDropdown(e, 'dataDropdownActive', false);
-      if (!isAboutBtn) this.toggleDataDropdown(e, 'aboutDropdownActive', false);
-    }
-  }
-
-  toggleDataDropdown(e, specificDropdown, to) {
+  // This function is debounced. If you don't do that insane things will happen
+  toggleDropdown(specificDropdown, to) {
     this.setState({
+      ...{ dataActive: false, aboutActive: false },
       [specificDropdown]: to
-    });
-
-    requestAnimationFrame(() => {
-      window[!to ?
-        'removeEventListener' : 'addEventListener']('click', this.onScreenClick, true);
     });
   }
 
   render() {
-    const dataDropDown = (
-      <TetherComponent
-        attachment="top right"
-        constraints={[{
-          to: 'window'
-        }]}
-        targetOffset="0px 100%"
-        classes={{
-          element: 'c-tooltip -arrow-right -menu'
-        }}
-      >
-        {/* First child: This is what the item will be tethered to */}
-        <a // eslint-disable-line jsx-a11y/no-static-element-interactions
-          ref={(c) => { this.dataDropdownBtn = c; }}
-          role="link"
-          onClick={e => this.toggleDataDropdown(e, 'dataDropdownActive', !this.state.dataDropdownActive)}
-        >
-          Data
-        </a>
-        {/* Second child: If present, this item will be tethered to the the first child */}
-        {this.state.dataDropdownActive &&
-          <ul className="data-dropdown">
-            <li>
-              <Link
-                route="explore"
-                onClick={e => this.toggleDataDropdown(e, 'dataDropdownActive', false)}
-              >
-                <a>Explore Datasets</a>
-              </Link>
-            </li>
-            <li>
-              <Link
-                route="dashboards"
-                onClick={e => this.toggleDataDropdown(e, 'dataDropdownActive', false)}
-              >
-                <a>Dashboards</a>
-              </Link>
-            </li>
-            <li>
-              <Link
-                route="pulse"
-                onClick={e => this.toggleDataDropdown(e, 'dataDropdownActive', false)}
-              >
-                <a>Planet Pulse</a>
-              </Link>
-            </li>
-            <li>
-              <Link
-                route="get_involved"
-                onClick={e => this.toggleDataDropdown(e, 'dataDropdownActive', false)}
-              >
-                <a>Explore Tools</a>
-              </Link>
-            </li>
-          </ul>
-        }
-      </TetherComponent>
-      );
-
-    const aboutDropDown = (
-      <TetherComponent
-        attachment="top right"
-        constraints={[{
-          to: 'window'
-        }]}
-        targetOffset="0px 100%"
-        classes={{
-          element: 'c-tooltip -arrow-right -menu'
-        }}
-      >
-        {/* First child: This is what the item will be tethered to */}
-        <a // eslint-disable-line jsx-a11y/no-static-element-interactions
-          ref={(c) => { this.aboutDropdownBtn = c; }}
-          role="link"
-          onClick={e => this.toggleDataDropdown(e, 'aboutDropdownActive', !this.state.aboutDropdownActive)}
-        >
-          About
-        </a>
-        {/* Second child: If present, this item will be tethered to the the first child */}
-        {this.state.aboutDropdownActive &&
-          <ul
-            className="data-dropdown"
-          >
-            <li>
-              <Link
-                route="about"
-                onClick={e => this.toggleDataDropdown(e, 'aboutDropdownActive', false)}
-              >
-                <a>About</a>
-              </Link>
-            </li>
-            <li>
-              <Link
-                route="about_partners"
-                onClick={e => this.toggleDataDropdown(e, 'aboutDropdownActive', false)}
-              >
-                <a>Partners</a>
-              </Link>
-            </li>
-          </ul>
-        }
-      </TetherComponent>
-      );
-
-    const navigationLinks = [
-      { name: dataDropDown },
-      { name: <a href="/insights">Insights</a> },
-      { name: aboutDropDown },
-      { name: <Link route="get_involved"><a>Get Involved</a></Link> }
+    const items = [
+      {
+        name: 'Data',
+        component: <HeaderDropdownData
+          active={this.state.dataActive}
+          onMouseEnter={() => this.toggleDropdown('dataActive', true)}
+          onMouseLeave={() => this.toggleDropdown('dataActive', false)}
+        />
+      },
+      {
+        name: 'Insights',
+        component: <Link route="insights"><a>Insights</a></Link>
+      },
+      {
+        name: 'About',
+        component: <HeaderDropdownAbout
+          active={this.state.aboutActive}
+          onMouseEnter={() => this.toggleDropdown('aboutActive', true)}
+          onMouseLeave={() => this.toggleDropdown('aboutActive', false)}
+        />
+      },
+      {
+        name: 'Get Involved',
+        component: <Link route="get_involved"><a>Get Involved</a></Link>
+      },
+      {
+        name: 'My RW',
+        component: <HeaderUser />
+      }
     ];
 
-    const mainClass = this.props.fullScreen ? '-fullScreen' : '';
+    const headerClass = classnames({
+      '-transparent': this.props.pageHeader
+    });
 
     return (
-      <header className="c-header">
+      <header className={`c-header ${headerClass}`}>
         <div className="header-secondary">
-          {/* Language selector */}
+          {/* We will load the script generated */}
         </div>
-        <div className={`header-main ${mainClass}`}>
-          <div className="brand">
-            <Link route="home"><a><Logo /></a></Link>
-          </div>
-          <div className="menu">
-            <Menu items={navigationLinks} />
+        <div className="l-container">
+          <div className="header-main">
+            <h1 className="header-logo">
+              <Link route="home">
+                <a>
+                  <svg><use xlinkHref="#icon-logo" /></svg>
+                  <span>Resource Watch</span>
+                </a>
+              </Link>
+            </h1>
+            <nav className="header-menu">
+              <ul>
+                {items.map(item => (
+                  <li key={item.name}>
+                    {item.component}
+                  </li>
+                ))}
+              </ul>
+            </nav>
           </div>
         </div>
       </header>
@@ -184,8 +103,13 @@ class Header extends React.Component {
 
 }
 
+Header.defaultProps = {
+  pageHeader: false
+};
+
+
 Header.propTypes = {
-  fullScreen: React.PropTypes.bool
+  pageHeader: PropTypes.bool
 };
 
 export default Header;
