@@ -1,25 +1,28 @@
 import React from 'react';
 import { Link } from 'routes';
+import Icon from 'components/ui/Icon';
 
 function Breadcrumbs(props) {
   return (
     <ul className="c-breadcrumbs">
-      {props.items.map((item, i) => {
-        if (!props.current) {
-          return <li key={i} className="item"><Link route={item.url}><a>{item.name}</a></Link></li>;
-        }
-        return i < (props.items.length - 1) ?
-          <li key={i} className="item"><Link route={item.url}><a>{item.name}</a></Link></li> :
-          <li key={i} className="item -current">{item.name}</li>;
-      })}
+      {props.items.map((item, i) => (
+        <li key={i} className="item">
+          <Link route={item.url}>
+            <a>
+              {props.items.length === 1 &&
+                <Icon className="c-icon -smaller" name="icon-arrow-left" />
+              }
+              <span>{item.name}</span>
+            </a>
+          </Link>
+        </li>
+      ))}
     </ul>
   );
 }
 
 Breadcrumbs.propTypes = {
-  items: React.PropTypes.array,
-  /* Current page present on breadcrumbs */
-  current: React.PropTypes.bool
+  items: React.PropTypes.array
 };
 
 Breadcrumbs.defaultProps = {
