@@ -1,36 +1,24 @@
 import React from 'react';
+
+// Redux
 import withRedux from 'next-redux-wrapper';
 import { initStore } from 'store';
-import classNames from 'classnames';
 import { resetDataset, toggleLayerShown } from 'redactions/exploreDetail';
 import updateLayersShown from 'selectors/explore/layersShownExploreDetail';
+
+// Next
+import { Link } from 'routes';
+
+// Services
+import DatasetService from 'services/DatasetService';
 
 // Components
 import Page from 'components/app/layout/Page';
 import Title from 'components/ui/Title';
 import Breadcrumbs from 'components/ui/Breadcrumbs';
-import Button from 'components/ui/Button';
-import Icon from 'components/ui/Icon';
-import DatasetList from 'components/app/explore/DatasetList';
 import Spinner from 'components/ui/Spinner';
-import Map from 'components/vis/Map';
-import Legend from 'components/ui/Legend';
-import LayerManager from 'utils/layers/LayerManager';
-import DatasetService from 'services/DatasetService';
 import WidgetEditor from 'components/widgets/WidgetEditor';
-
-const breadcrumbs = [
-  { name: 'Home', url: 'home' },
-  { name: 'Explore', url: 'explore' }
-];
-
-const mapConfig = {
-  zoom: 3,
-  latLng: {
-    lat: 0,
-    lng: 0
-  }
-};
+// import DatasetList from 'components/app/explore/DatasetList';
 
 class ExploreDetail extends React.Component {
 
@@ -45,8 +33,7 @@ class ExploreDetail extends React.Component {
     this.state = {
       similarDatasetsLoaded: false,
       dataset: null,
-      loading: false,
-      layers: []
+      loading: false
     };
 
     // DatasetService
@@ -55,6 +42,12 @@ class ExploreDetail extends React.Component {
     });
   }
 
+  /**
+   * Component Lifecycle
+   * - componentDidMount
+   * - componentWillReceiveProps
+   * - componentWillUnmount
+  */
   componentDidMount() {
     this.getDataset();
   }
@@ -76,6 +69,10 @@ class ExploreDetail extends React.Component {
     this.props.resetDataset();
   }
 
+  /**
+   * HELPERS
+   * - getDataset
+  */
   getDataset() {
     this.setState({
       loading: true
@@ -87,7 +84,7 @@ class ExploreDetail extends React.Component {
           loading: false
         });
       }).catch((error) => {
-        console.log(error);
+        console.error(error);
         this.setState({
           loading: false
         });
@@ -95,99 +92,99 @@ class ExploreDetail extends React.Component {
     });
   }
 
+  /**
+   * UI EVENTS
+   * - triggerDownload
+  */
   triggerDownload() {
     console.info('triggerDownload');
   }
 
   render() {
-    const { dataset, loading, layers } = this.state;
+    const { dataset, loading } = this.state;
     const metadata = dataset && dataset.attributes.metadata;
-
-    // const similarDatasetsSectionClass = classNames({
-    //   row: true,
-    //   'similar-datasets-row': true,
-    //   '-active': exploreDetail.similarDatasets.list.filter(value =>
-    //               value.id !== this.props.datasetID
-    //             ).length > 0
-    // });
-
-    const pageStructure = (
-      <div className="c-page c-page-explore-detail">
-        <div className="row">
-          <div className="column small-12">
-            <Breadcrumbs items={breadcrumbs} />
-            <Title className="-primary -huge title" >{ dataset && dataset.attributes &&
-                dataset.attributes.name}</Title>
-          </div>
-        </div>
-        <div className="row widget-row">
-          <div className="column small-12 ">
-            {dataset &&
-              <WidgetEditor
-                dataset={dataset.id}
-              />
-            }
-            <Spinner
-              isLoading={loading}
-              className="-light"
-            />
-          </div>
-        </div>
-        <div className="row description-row">
-          <div className="column small-2 social" >
-            <Icon name="icon-twitter" className="-small" />
-            <Icon name="icon-facebook" className="-small" />
-          </div>
-          <div className="column small-7">
-            {/* Description */}
-            {metadata && (metadata.length > 0)
-              && metadata[0].attributes.description &&
-              <p>{metadata[0].attributes.description}</p>
-            }
-          </div>
-          <div className="column small-3 actions">
-            <Button
-              properties={{
-                disabled: true,
-                className: '-primary -fullwidth -disabled'
-              }}
-              onClick={this.triggerDownload}
-            >
-              Download
-            </Button>
-          </div>
-        </div>
-      </div>
-    );
-
-    // <div>
-    //   <div className="column small-12">
-    //     <Title className="-secondary title">
-    //       Similar datasets
-    //     </Title>
-    //   </div>
-    //   <div className="column small-12">
-    //     <DatasetList
-    //       active={exploreDetail.similarDatasets.list.map(value => value.id)}
-    //       list={exploreDetail.similarDatasets.list.filter(value =>
-    //         value.id !== this.props.datasetID
-    //       )}
-    //       mode="grid"
-    //     />
-    //     <Spinner
-    //       isLoading={exploreDetail.similarDatasets.loading}
-    //       className="-relative"
-    //     />
-    //   </div>
-    // </div>
 
     return (
       <Page
         title="Explore detail"
         description="Explore detail description..."
+        pageHeader
       >
         <div className="c-page-explore-detail">
-          {pageStructure}
+          <Spinner
+            isLoading={loading}
+            className="-fixed -light"
+          />
+
+          {/* PAGE HEADER */}
+          <div className="c-page-header">
+            <div className="l-container">
+              <div className="page-header-content -padding-b-2">
+                <Breadcrumbs
+                  items={[{ name: 'Explore datasets', url: 'explore' }]}
+                />
+
+                <Title className="-primary -huge page-header-title" >
+                  { dataset && dataset.attributes && dataset.attributes.name}
+                </Title>
+
+                <div className="page-header-info">
+                  <ul>
+                    <li>Source: {(metadata && metadata.length > 0 && metadata[0].source) || '-'}</li>
+                    <li>Last update: {dataset && dataset.attributes && new Date(dataset.attributes.updatedAt).toJSON().slice(0, 10).replace(/-/g, '/')}</li>
+                    {/* Favorites <li>Last update: {dataset && dataset.attributes && dataset.attributes.updatedAt}</li> */}
+                  </ul>
+                </div>
+              </div>
+            </div>
+          </div>
+
+          {/* WIDGET EDITOR */}
+          {dataset &&
+            <WidgetEditor
+              dataset={dataset.id}
+            />
+          }
+
+          {/* DATASET INFO && ACTIONS */}
+          <div className="c-page-section">
+            <section className="c-dataset-info">
+              <div className="row">
+                <div className="column small-12 medium-7">
+                  {/* Description */}
+                  <div className="dataset-info-description">
+                    {metadata && (metadata.length > 0) && metadata[0].attributes.description &&
+                      metadata[0].attributes.description
+                    }
+                  </div>
+                </div>
+                <div className="column large-offset-2 small-3">
+                  <div className="dataset-info-actions">
+                    <div className="row flex-dir-column">
+                      <div className="column">
+                        <button
+                          disabled
+                          className="c-button -primary -fullwidth -disabled"
+                          onClick={this.triggerDownload}
+                        >
+                          Download
+                        </button>
+                      </div>
+                    </div>
+                  </div>
+                </div>
+              </div>
+            </section>
+          </div>
+
+
+          {/* RELATED TOOLS */}
+
+          {/* SIMILAR DATASETS */}
+
+          {/* RELATED INSIGHTS */}
+
+          {/* PLANET PULSE */}
         </div>
       </Page>
     );
